requirement.txt
python-docx==1.2.0
pyttsx3==2.99